[sandbox] Introduce Code::parameter_count
This field is currently only available for optimized code, in which case
it contains the parameter count of the compiled function, similar to
BytecodeArray::parameter_count. For now, this (trusted) parameter count
is only used during deoptimization where we were previously using the
(untrusted) parameter count stored on the SharedFunctionInfo, which
could lead to various issues if the value was modified by an attacker.

Bug: chromium:335544065
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5465515
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#93462}

+++ src/interpreter/bytecode-array-writer.cc
@@ -39,7 +39,7 @@ BytecodeArrayWriter::BytecodeArrayWriter(
 
 template <typename IsolateT>
 Handle<BytecodeArray> BytecodeArrayWriter::ToBytecodeArray(
-    IsolateT* isolate, int register_count, int parameter_count,
+    IsolateT* isolate, int register_count, uint16_t parameter_count,
     Handle<TrustedByteArray> handler_table) {
   DCHECK_EQ(0, unbound_jumps_);
 
@@ -55,11 +55,11 @@ Handle<BytecodeArray> BytecodeArrayWriter::ToBytecodeArray(
 
 template EXPORT_TEMPLATE_DEFINE(V8_EXPORT_PRIVATE)
     Handle<BytecodeArray> BytecodeArrayWriter::ToBytecodeArray(
-        Isolate* isolate, int register_count, int parameter_count,
+        Isolate* isolate, int register_count, uint16_t parameter_count,
         Handle<TrustedByteArray> handler_table);
 template EXPORT_TEMPLATE_DEFINE(V8_EXPORT_PRIVATE)
     Handle<BytecodeArray> BytecodeArrayWriter::ToBytecodeArray(
-        LocalIsolate* isolate, int register_count, int parameter_count,
+        LocalIsolate* isolate, int register_count, uint16_t parameter_count,
         Handle<TrustedByteArray> handler_table);
 
 template <typename IsolateT>
