+++ test/cctest/heap/test-heap.cc
@@ -625,7 +625,7 @@ TEST(BytecodeArray) {
   static const uint8_t kRawBytes[] = {0xC3, 0x7E, 0xA5, 0x5A};
   static const int kRawBytesSize = sizeof(kRawBytes);
   static const int32_t kFrameSize = 32;
-  static const int32_t kParameterCount = 2;
+  static const uint16_t kParameterCount = 2;
 
   ManualGCScope manual_gc_scope;
   heap::ManualEvacuationCandidatesSelectionScope
